just output template
A project to list files of a directory in an html file. This one only prints out the surrounding template I intend to begin with.

Next: get just this folder's name, list the oswalk on the command line, put them naively in the file, and then make unordered lists with them.