Works as advertised
This now checks and emits links to arbitrary folder depths. (Just checked that with a less trivial example.) I can't promise what happens to the <h_> tags when you go below 5, though.

Also added a note that this is mit license. To be honest, that's half the reason I bothered. I found another python version, but it was gpl. The other half was that it depended on other libraries.